+++ src/impls/sets/btree_set.rs
@@ -1,4 +1,5 @@
 use alloc::collections::BTreeSet;
+use fluent_result::bool::dbg::Expect;
 
 crate::impls::macros::impl_try_from_iter_via_try_extend_one!(
     type: BTreeSet<T> where [T: Ord] of T;
@@ -8,17 +9,17 @@ crate::impls::macros::impl_try_from_iter_via_try_extend_one!(
 crate::impls::macros::impl_try_extend_via_try_extend_one!(
     type: BTreeSet<T> where [T: Ord] of T;
     reserve: |_, _| ();
-    build_empty_collection: |_| BTreeSet::new()
+    build_empty: |_| BTreeSet::new()
 );
 
 crate::impls::macros::impl_try_extend_safe_for_colliding_type!(
     type: BTreeSet<T> where [T: Ord] of T;
     build_staging: |_, _| BTreeSet::new();
-    contains: |set, item| set.contains(item)
+    contains: BTreeSet::contains
 );
 
 crate::impls::macros::impl_try_extend_one_for_colliding_type!(
     type: BTreeSet<T> where [T: Ord] of T;
-    contains: |set, item| set.contains(item);
-    insert: |set, item| { set.insert(item); }
+    contains: BTreeSet::contains;
+    insert: |set, item| set.insert(item).expect_true("insert should succeed after contains check")
 );